changes v2

+++ src/main/java/com/io/admin/controller/ApplicationController.java
@@ -1,8 +1,6 @@
 package com.io.admin.controller;
 
-import com.io.admin.model.Administrator;
-import com.io.admin.model.Member;
-import com.io.admin.model.Work;
+import com.io.admin.model.*;
 import com.io.admin.service.ApplicationService;
 import com.io.admin.exception.BusinessException;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -10,6 +8,8 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+
+import java.util.List;
 import java.util.Optional;
 
 @RestController
@@ -34,6 +34,8 @@ public class ApplicationController {
     //################  BASE DE DATOS ADMIN.IO  ###################
     //#############################################################
 
+    //TODO CREATE
+
     //CREAR MEMBERS
     @RequestMapping(value = "/createMember", method = RequestMethod.POST)
     public ResponseEntity<Member> create(@RequestBody Member member){
@@ -52,13 +54,26 @@ public class ApplicationController {
         Work work1 = applicationService.addWork(work);
         return new ResponseEntity(work, HttpStatus.CREATED);
     }
+    //CREAR WORKGROUP
+    @RequestMapping(value = "/createWorkgroup", method = RequestMethod.POST)
+    public ResponseEntity<Workgroup> addWorkgroup(@RequestBody Workgroup workgroup){
+        Workgroup workgroup2 = applicationService.addWorkgroup(workgroup);
+        return new ResponseEntity(workgroup, HttpStatus.OK);
+    }
+    //CREAR RELACION WORK AND GROUP
 
-    //################ READ ########################################
+
+    //TODO READ
     //member
     @RequestMapping(value = "/getMember", method = RequestMethod.GET)
     public ResponseEntity<Member> userById(@RequestParam(value = "id") long id) {
         Optional<Member> user = applicationService.getMember(id);
         return new ResponseEntity(user, HttpStatus.OK);
+    }
+    @RequestMapping(value = "/getMembers", method = RequestMethod.GET)
+    public ResponseEntity<Member> memberList(){
+        List<Member> members = applicationService.memberlist();
+        return new ResponseEntity(members, HttpStatus.OK);
     }
     //admin
     @RequestMapping(value = "/getAdmin", method = RequestMethod.GET)
@@ -66,17 +81,81 @@ public class ApplicationController {
         Optional<Administrator> admin = applicationService.getAdmin(id);
         return new ResponseEntity(admin, HttpStatus.OK);
     }
+    @RequestMapping(value="/getAdmins", method = RequestMethod.GET)
+    public ResponseEntity<Administrator> adminlist(){
+        List<Administrator> administrators = applicationService.adminlist();
+        return new ResponseEntity(administrators, HttpStatus.OK);
+    }
     //work
     @RequestMapping(value = "/getWork", method = RequestMethod.GET)
     public ResponseEntity<Work> workById(@RequestParam(value = "id") long workid){
         Optional<Work> work = applicationService.getWork(workid);
         return new ResponseEntity(work, HttpStatus.OK);
     }
+    @RequestMapping(value = "/getWorks", method = RequestMethod.GET)
+    public ResponseEntity<Work> worklist(){
+        List<Work> works = applicationService.worklist();
+        return new ResponseEntity(works, HttpStatus.OK);
+    }
+    //workgroups
+    @RequestMapping(value = "/getWorkgroup", method = RequestMethod.GET)
+    public ResponseEntity<Workgroup> workgroupById(@RequestParam(value = "id") long workgroupid){
+        Optional<Workgroup> work = applicationService.getWorkgroup((int)workgroupid);
+        return new ResponseEntity(work, HttpStatus.OK);
+    }
+    @RequestMapping(value = "/getWorkgroups", method = RequestMethod.GET)
+    public ResponseEntity<Workgroup> workgrouplist(){
+        List<Workgroup> workgroups = applicationService.workgroupList();
+        return new ResponseEntity(workgroups, HttpStatus.OK);
+    }
 
-    //################### DELETE
+    //TODO UPDATE
+
+
+    //TODO DELETE
     @RequestMapping(value = "/deleteMember")
     public String deleteMember(@RequestParam(value = "id") int id){
-        applicationService.deleteMember(id);
-        return "borrado";
+        Optional<Member> m1 = applicationService.getMember(id);
+        System.out.println(m1);
+        if(m1.isEmpty()){
+            return "no existe";
+        }
+        else{
+            applicationService.deleteMember(id);
+            return "borrado";
+        }
+    }
+    @RequestMapping(value = "/deleteAdmin")
+    public String deleteAdmin(@RequestParam(value = "id") int id){
+        Optional<Administrator> administrator = applicationService.getAdmin(id);
+        if (administrator.isEmpty()){
+            return "no existe";
+        }
+        else {
+            applicationService.deleteAdmin(id);
+            return "Borrado";
+        }
+    }
+    @RequestMapping(value = "/deleteWork")
+    public String deleteWork(@RequestParam(value = "id") int id){
+        Optional<Work> work = applicationService.getWork(id);
+        if (work.isEmpty()){
+            return "No existe";
+        }
+        else {
+            applicationService.deleteWork(id);
+            return "Borrado";
+        }
+    }
+    @RequestMapping(value = "/deleteWorkgroup")
+    public String deleteWorkgroup(@RequestParam(value = "id")int id){
+        Optional<Workgroup> workgroup = applicationService.getWorkgroup(id);
+        if (workgroup.isEmpty()){
+            return "No existe";
+        }
+        else {
+            applicationService.deleteWorkgroup(id);
+            return "Borrado";
+        }
     }
 }
